Modified accessor variable name key by slice Add a second test where min and max are multidimensional
